astronauts-spark-launcher/launcher/requirements.txt
fastapi
uvicorn
kubernetes
PyYAML
requests